Changes:  * moving towards centralised build-system, using svn:externals for dependencies (FLAC uses CVS, hence it's libs/frameworks build files defined locally);  * Xcode build files now produce XiphQT as a single (statically linked) bundle (full, or decoders-only flavors);  * Theora decoder: fixed handling of multiple-of-16 frame sizes and offsets (leftover from transition to theora-exp);  * Ogg importer: added FLAC channel mapping layouts for 2.1 - 5.1;  * Ogg importer: initial attempt to use AddMediaSampleReferences64 bulk update interface;  * Theora encoder: initial version;  * Ogg exporter: initial support for video streams.
(The revision might not compile or work properly.)

+++ xiph-qt/CAFLAC/src/CAFLACDecoder.cpp
@@ -536,7 +536,7 @@ void CAFLACDecoder::DFPclear()
 
 #pragma mark Callbacks
 
-::FLAC__StreamDecoderReadStatus CAFLACDecoder::read_callback(FLAC__byte buffer[], unsigned *bytes)
+::FLAC__StreamDecoderReadStatus CAFLACDecoder::read_callback(FLAC__byte buffer[], size_t *bytes)
 {
     dbg_printf(" | -> [%08lx] :: read_callback(%ld)\n", (UInt32) this, *bytes);
     FLAC__StreamDecoderReadStatus ret = FLAC__STREAM_DECODER_READ_STATUS_CONTINUE;
